Выводит параметры дополнительных предложений
